Fix for (most incidences of) randomly removed clothes.
If _clothingState is uninitialized, the first call to put clothes back on (after changing clothes) wouldn't actually call UpdateClothingState and modify the controller's chaCtrl. Thus, when it later saved the data, it saved the undressed state instead of the intended clean state. I extended this to at least accessory/siru, not sure if it applies to the other states.

Minor: Post showers didn't force update clothing state. Probably was not a major issue.